requirements.txt
flask==2.0.1
werkzeug==2.0.1
mpi4py==3.1.3
numpy==1.21.0
pillow==9.0.0
matplotlib==3.4.2 